fracture: split admin settings payment and community panels
